% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeral.R
\name{numeral}
\alias{numeral}
\title{Numeral class}
\usage{
numeral(x = numeric(), system = c("en", "ar", "fa"))
}
\arguments{
\item{x}{A numeric vector.}

\item{system}{Two-letter language code of the desired numeral system; see
details for a list of available systems. Default: \code{"en"}.}
}
\value{
Vector of class \code{numeral}.
}
\description{
The \code{numeral} class extends the base numeric vectors with methods for
printing them using UTF digits from other numeral systems.
}
\details{
The following numeral systems are currently supported:
\itemize{
\item \code{"en"}: Western Arabic numerals (the default display in base R)
\item \code{"ar"}: Eastern Arabic numerals
\item \code{"fa"}: Persian numerals
}

Decimal separators etc. are not currently localised.
}
\examples{
# Eastern Arabic numerals
numeral(1:10, "ar")

# Persian numerals
numeral(1:10, "fa")
}
