% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerals-package.R
\docType{package}
\name{numerals-package}
\alias{numerals}
\alias{numerals-package}
\title{numerals: Numeral Systems}
\description{
Provides an S3 vector class ('numeral') representing numerics in
   different numeral systems.
}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})

}
\keyword{internal}
