% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numr_replace.R
\name{numr_replace}
\alias{numr_replace}
\title{Replace numerals in a string}
\usage{
numr_replace(x, system)
}
\arguments{
\item{x}{Character. Input vector.}

\item{system}{Two-letter language code of the replacement numeral system.}
}
\value{
Character vector the same length as \code{x} with numerals replaced.
}
\description{
\code{numr_replace()} replaces the numerals in a string with another numeral
system. It is vectorised over \code{x}.
}
\examples{
numr_replace("0123456789", "ar")
}
