% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeral.R
\name{numr_system}
\alias{numr_system}
\title{Get or set the numeral system of a vector}
\usage{
numr_system(x)
}
\arguments{
\item{x}{\link{numeral} vector.}
}
\description{
These functions retrieve or replace the \code{system} attribute of a \link{numeral}
vector.
}
\examples{
x <- numeral(1, "ar")
numr_system(x)
}
