% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerals-package.R
\docType{package}
\name{numerals-package}
\alias{numerals}
\alias{numerals-package}
\title{numerals: UTF Numeral Systems}
\description{
Extends the base numeric types with methods for printing using UTF
   digits from international numeral systems. Currently supports Eastern Arabic
   ('ar'), Bengali ('bn'), Persian ('fa'), and Burmese ('my') numerals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joeroe/numerals}
  \item Report bugs at \url{https://github.com/joeroe/numerals/issues}
}

}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})

}
\keyword{internal}
